+++ common/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 This document contains all changes to the crate since version 0.9.4.
 
+## 0.13.0
+
+### Breaking changes
+
+- Renamed `Error` to `EncodeError`.
+- Made `zalgo_decode` return a `DecodeError` instead of a `FromUtf8Error`.
+
+### Other changes
+
+- Added the `DecodeError` error type.
+- Fixed bug in `zalgo_decode` that made it attempt to allocate
+ `usize::MAX/2` bytes if the input string was empty.
+
 ## 0.12.2
 
 - Documentation improvements.